feat: bypass cors also for binary files
