+++ main.ts
@@ -1,8 +1,42 @@
+import * as cors from 'cors';
 import { BrowserWindow, Menu, MenuItem, app, nativeImage, shell } from 'electron';
 import * as settings from 'electron-settings';
+import * as express from 'express';
+import * as morgan from 'morgan';
 import * as path from 'path';
 import * as url from 'url';
-import { appRest } from './cors';
+
+// Inspired by https://github.com/Miorey/bypass-cors-policies
+export const appRest = express();
+const TIMEOUT = 2000;
+const SERVER_NAME = 'https://wow.zamimg.com';
+const fetch = require(`node-fetch`);
+
+appRest.use(morgan(`dev`));
+appRest.use(express.json());
+appRest.use(express.urlencoded({ extended: true }));
+appRest.use(cors());
+
+async function getFile(req) {
+  console.log(`${SERVER_NAME}${req.url}`);
+  const timeout = new Promise((_, reject) => {
+    setTimeout(() => {
+      reject(new Error(`Request timed out after ${TIMEOUT} milli-seconds`));
+    }, TIMEOUT);
+  });
+
+  const response: any = await Promise.race([fetch(`${SERVER_NAME}${req.url}`), timeout]).catch(() => ({ ok: false, status: 408 }));
+  if (!response.ok) {
+    return response.status;
+  }
+  const buffer = await response.arrayBuffer();
+  return Buffer.from(buffer);
+}
+
+appRest.use(`/`, async (req, res, next) => {
+  const content = await getFile(req);
+  res.send(content);
+});
 
 let win;
 const args = process.argv.slice(1);
